requirements.txt
streamlit==1.37.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
gTTS==2.5.3
reportlab==4.2.2
Pillow==10.4.0
groq==0.11.0
httpx<0.28